# packages/api-client-generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FPS Control Plane",
    "description": "Canonical HTTP contract for the control plane. Generated from source.",
    "license": {
      "name": "LicenseRef-UNDECIDED",
      "identifier": "LicenseRef-UNDECIDED"
    },
    "version": "0.0.1-alpha.1"
  },
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/auth/totp/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "confirm_totp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpConfirmResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/invitations": {
      "get": {
        "tags": [
          "identity"
        ],
        "operationId": "list_invitations",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "identity"
        ],
        "operationId": "create_invitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationView"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invitations/accept": {
      "post": {
        "tags": [
          "identity"
        ],
        "operationId": "accept_invitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/nodes/enroll": {
      "post": {
        "tags": [
          "nodes"
        ],
        "operationId": "enroll",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollResponse"
                }
              }
            }
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/v1/nodes/enrollment-tokens": {
      "post": {
        "tags": [
          "nodes"
        ],
        "operationId": "create_enrollment_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nodes/{id}": {
      "get": {
        "tags": [
          "nodes"
        ],
        "operationId": "get_node",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/v1/nodes/{id}/heartbeat": {
      "post": {
        "tags": [
          "nodes"
        ],
        "operationId": "heartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/nodes/{id}/revoke": {
      "post": {
        "tags": [
          "nodes"
        ],
        "operationId": "revoke_node",
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/schedules": {
      "get": {
        "tags": [
          "servers"
        ],
        "operationId": "list_schedules",
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "servers"
        ],
        "operationId": "create_schedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/servers": {
      "get": {
        "tags": [
          "servers"
        ],
        "operationId": "list_servers",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "servers"
        ],
        "operationId": "create_server",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSummary"
                }
              }
            }
          }
        }
      }
    },
    "/v1/servers/{id}": {
      "get": {
        "tags": [
          "servers"
        ],
        "operationId": "get_server",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "servers"
        ],
        "operationId": "delete_server",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSummary"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "servers"
        ],
        "operationId": "patch_server",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/settings": {
      "get": {
        "tags": [
          "ops"
        ],
        "operationId": "get_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSettingsView"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ops"
        ],
        "operationId": "patch_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSettingsView"
                }
              }
            }
          }
        }
      }
    },
    "/v1/setup": {
      "post": {
        "tags": [
          "setup"
        ],
        "operationId": "complete_setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "409": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "operationId": "list_templates",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "templates"
        ],
        "operationId": "create_template",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSummary"
                }
              }
            }
          }
        }
      }
    },
    "/v1/templates/import-egg": {
      "post": {
        "tags": [
          "templates"
        ],
        "operationId": "import_egg_template",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSummary"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users/{id}": {
      "patch": {
        "tags": [
          "identity"
        ],
        "operationId": "patch_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInvitationRequest": {
        "type": "object",
        "required": [
          "token",
          "password",
          "display_name"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          }
        }
      },
...
      "CreateInvitationRequest": {
        "type": "object",
        "required": [
          "email",
          "role"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "EnrollmentTokenRequest": {
        "type": "object",
        "properties": {
          "label": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "EnrollmentTokenResponse": {
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InvitationId": {
        "type": "string",
        "format": "uuid"
      },
      "InvitationView": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "expires_at"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/InvitationId"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "accepted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "token": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "totp_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "recovery_code": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "user",
          "permissions",
          "access_token",
          "refresh_token",
          "csrf_token",
          "expires_in",
          "mfa_required"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "csrf_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "mfa_required": {
            "type": "boolean"
          },
          "mfa_token": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PatchSettingsRequest": {
        "type": "object",
        "properties": {
          "operator_notes": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TotpConfirmRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "TotpConfirmResponse": {
        "type": "object",
        "required": [
          "recovery_codes"
        ],
        "properties": {
          "recovery_codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}